Move 2 components wrapped in loadData to containers
See
https://github.com/MoveOnOrg/Spoke/pull/1293#pullrequestreview-333327222
